Optimize spec page loading everywhere
Build currently fails with a timeout when it tries to load the ED of CSS2 to
extract the list of pages from the table of contents (the ED is not a multipage
spec per se, but the /TR version is, and the "multipage" flag applies to both).

That timeout is due to images taking a long time to load on the CSS drafts
server right now for some reason. Puppeteer surrenders after 30s of trying to
fetch the images. That is not specific to the context under which the build is
run: I get the same timeout when I run things locally.

As a workaround, this update reuses in extract-pages the same load logic as
the one used in fetch-info, which itself is a copy-and-paste from the logic
used in Reffy to crawl the specs. That logic notably avoids loading resources
identified as images, which solves the timeout issue.

This also has the advantage of making us use a consistent approach whenever we
need to load a spec using Puppeteer.

A side update: the fetch-info logic was always sending a GitHub API request to
retrieve IETF HTTPWG repository info, even when that was clearly not necessary.
To avoid hitting rate limits too fast, the request is now only sent when
absolutely necessary.
